+++ src/tests/Equinor.ProCoSys.IPO.WebApi.IntegrationTests/Invitations/ParticipantDtoGet.cs
@@ -10,13 +10,4 @@ namespace Equinor.ProCoSys.IPO.WebApi.IntegrationTests.Invitations
         public InvitedPersonDto Person { get; set; }
         public FunctionalRoleDto FunctionalRole { get; set; }
     }
-
-    public class ParticipantDtoEdit
-    {
-        public Organization Organization { get; set; }
-        public int SortKey { get; set; }
-        public ExternalEmailDto ExternalEmail { get; set; }
-        public PersonDto Person { get; set; }
-        public FunctionalRoleDto FunctionalRole { get; set; }
-    }
 }
