+++ src/tests/Equinor.ProCoSys.IPO.WebApi.IntegrationTests/KnownTestData.cs
@@ -10,7 +10,6 @@ namespace Equinor.ProCoSys.IPO.WebApi.IntegrationTests
         public static string ProjectDescription => "Test - Project";
         public static string McPkgNo => "MC10-2034";
         public static string CommPkgNo => "COMM12-2387";
-        public static string DisciplineCode => "A";
         public static string InvitationTitle => "TestInvitation";
         public static string InvitationDescription => "Test - Invitation";
         public static Guid MeetingId => new Guid("{818E6882-A5F1-4367-B459-1A2E1EE01D7F}");
